Add owned clients for each supported HTTP client
Add a wrapping Client for each supported HTTP client which owns the
token in use as well as its own HTTP client.

This brings the benefit of being slightly simpler to use (not needing to
supply the key for every request and keeping it readily available),
while trading off performance (it owns a Client instead of re-using one
via traits).